Preserve LTX2 HQ res2s noise precision

+++ python/sglang/multimodal_gen/runtime/pipelines_core/stages/ltx_2_denoising.py
@@ -556,10 +556,8 @@ class LTX2DenoisingStage(DenoisingStage):
                         dtype=sliced.dtype,
                     )
                     sliced = torch.cat([sliced, pad], dim=1)
-                return sliced.to(dtype=reference_tensor.dtype)
-        return cls._ltx2_res2s_new_noise(reference_tensor, generator).to(
-            dtype=reference_tensor.dtype
-        )
+                return sliced
+        return cls._ltx2_res2s_new_noise(reference_tensor, generator)
 
     @staticmethod
     def _ltx2_apply_clean_latent_mask(
@@ -740,14 +738,14 @@ class LTX2DenoisingStage(DenoisingStage):
         sub_noise_video = (
             self._ltx2_res2s_noise_like(
                 ctx.latents, ctx, substep=True, batch=batch
-            ).float()
+            )
             if ctx.use_native_hq_res2s_sde_noise
             else self._randn_like_with_batch_generators(ctx.latents, batch).float()
         )
         sub_noise_audio = (
             self._ltx2_res2s_noise_like(
                 ctx.audio_latents, ctx, substep=True, batch=batch, is_audio=True
-            ).float()
+            )
             if ctx.use_native_hq_res2s_sde_noise
             else self._randn_like_with_batch_generators(
                 ctx.audio_latents, batch
@@ -802,14 +800,14 @@ class LTX2DenoisingStage(DenoisingStage):
         step_noise_video = (
             self._ltx2_res2s_noise_like(
                 ctx.latents, ctx, substep=False, batch=batch
-            ).float()
+            )
             if ctx.use_native_hq_res2s_sde_noise
             else self._randn_like_with_batch_generators(ctx.latents, batch).float()
         )
         step_noise_audio = (
             self._ltx2_res2s_noise_like(
                 ctx.audio_latents, ctx, substep=False, batch=batch, is_audio=True
-            ).float()
+            )
             if ctx.use_native_hq_res2s_sde_noise
             else self._randn_like_with_batch_generators(
                 ctx.audio_latents, batch
@@ -2354,7 +2352,7 @@ class LTX2DenoisingStage(DenoisingStage):
                 substep_video_noise = (
                     self._ltx2_res2s_noise_like(
                         ctx.latents, ctx, substep=True, batch=batch
-                    ).float()
+                    )
                     if ctx.use_native_hq_res2s_sde_noise
                     else self._randn_like_with_batch_generators(
                         ctx.latents, batch
@@ -2367,7 +2365,7 @@ class LTX2DenoisingStage(DenoisingStage):
                         substep=True,
                         batch=batch,
                         is_audio=True,
-                    ).float()
+                    )
                     if ctx.use_native_hq_res2s_sde_noise
                     else self._randn_like_with_batch_generators(
                         ctx.audio_latents, batch
@@ -2429,7 +2427,7 @@ class LTX2DenoisingStage(DenoisingStage):
                 step_video_noise = (
                     self._ltx2_res2s_noise_like(
                         ctx.latents, ctx, substep=False, batch=batch
-                    ).float()
+                    )
                     if ctx.use_native_hq_res2s_sde_noise
                     else self._randn_like_with_batch_generators(
                         ctx.latents, batch
@@ -2442,7 +2440,7 @@ class LTX2DenoisingStage(DenoisingStage):
                         substep=False,
                         batch=batch,
                         is_audio=True,
-                    ).float()
+                    )
                     if ctx.use_native_hq_res2s_sde_noise
                     else self._randn_like_with_batch_generators(
                         ctx.audio_latents, batch
